+++ src/main/java/com/mytiki/account/features/latest/api_key/ApiKeyController.java
@@ -5,6 +5,7 @@
 
 package com.mytiki.account.features.latest.api_key;
 
+import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.features.latest.app_info.AppInfoService;
 import com.mytiki.account.utilities.Constants;
 import io.swagger.v3.oas.annotations.Operation;
@@ -22,6 +23,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+@XRayEnabled
 @Tag(name = "")
 @RestController
 @RequestMapping(value = Constants.API_LATEST_ROUTE)
