+++ src/main/java/com/mytiki/account/features/latest/api_key/ApiKeyService.java
@@ -5,7 +5,6 @@
 
 package com.mytiki.account.features.latest.api_key;
 
-import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.features.latest.app_info.AppInfoDO;
 import com.mytiki.account.features.latest.app_info.AppInfoService;
 import com.mytiki.account.features.latest.refresh.RefreshService;
@@ -34,7 +33,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@XRayEnabled
 public class ApiKeyService {
     private final ApiKeyRepository repository;
     private final AppInfoService appInfoService;
