+++ src/main/java/com/mytiki/account/features/latest/app_info/AppInfoController.java
@@ -5,6 +5,7 @@
 
 package com.mytiki.account.features.latest.app_info;
 
+import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.utilities.Constants;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
@@ -13,6 +14,7 @@ import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationToken;
 import org.springframework.web.bind.annotation.*;
 
+@XRayEnabled
 @Tag(name = "")
 @RestController
 @RequestMapping(value = AppInfoController.PATH_CONTROLLER)
