+++ src/main/java/com/mytiki/account/features/latest/exchange/ExchangeService.java
@@ -5,7 +5,6 @@
 
 package com.mytiki.account.features.latest.exchange;
 
-import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.features.latest.exchange.shopify.ShopifyClient;
 import com.mytiki.account.features.latest.refresh.RefreshService;
 import com.mytiki.account.features.latest.user_info.UserInfoAO;
@@ -22,7 +21,6 @@ import org.springframework.security.oauth2.core.OAuth2Error;
 import org.springframework.security.oauth2.core.OAuth2ErrorCodes;
 import org.springframework.security.oauth2.core.endpoint.OAuth2AccessTokenResponse;
 
-@XRayEnabled
 public class ExchangeService {
 
     private final UserInfoService userInfoService;
