+++ src/main/java/com/mytiki/account/features/latest/jwks/JwksController.java
@@ -5,6 +5,7 @@
 
 package com.mytiki.account.features.latest.jwks;
 
+import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.health.HealthController;
 import com.nimbusds.jose.jwk.JWKSet;
 import io.swagger.v3.oas.annotations.Operation;
@@ -14,6 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.Map;
 
+@XRayEnabled
 @RestController
 @RequestMapping(value = JwksController.ROUTE)
 public class JwksController {
