+++ src/main/java/com/mytiki/account/features/latest/org_info/OrgInfoService.java
@@ -5,7 +5,6 @@
 
 package com.mytiki.account.features.latest.org_info;
 
-import com.amazonaws.xray.spring.aop.XRayEnabled;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.ZonedDateTime;
@@ -13,7 +12,6 @@ import java.util.Optional;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-@XRayEnabled
 public class OrgInfoService {
     private final OrgInfoRepository repository;
