+++ src/main/java/com/mytiki/account/features/latest/refresh/RefreshService.java
@@ -5,7 +5,6 @@
 
 package com.mytiki.account.features.latest.refresh;
 
-import com.amazonaws.xray.spring.aop.XRayEnabled;
 import com.mytiki.account.security.oauth.OauthSub;
 import com.mytiki.account.utilities.Constants;
 import com.mytiki.account.utilities.builder.JwtBuilder;
@@ -25,7 +24,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
-@XRayEnabled
 public class RefreshService {
     private final RefreshRepository repository;
     private final JWSSigner jwtSigner;
