+++ src/main/java/com/mytiki/account/health/HealthController.java
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-@XRayEnabled
 @RestController
 @RequestMapping(value = HealthController.ROUTE)
 public class HealthController {
