+++ src/main/java/com/mytiki/account/utilities/XRayInspector.java
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) TIKI Inc.
+ * MIT license. See LICENSE file in root directory.
+ */
+
+package com.mytiki.account.utilities;
+
+import com.amazonaws.xray.AWSXRay;
+import com.amazonaws.xray.entities.Subsegment;
+import com.amazonaws.xray.spring.aop.AbstractXRayInterceptor;
+import com.amazonaws.xray.spring.aop.XRayInterceptorUtils;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.aspectj.lang.annotation.Aspect;
+import org.aspectj.lang.annotation.Pointcut;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.lang.invoke.MethodHandles;
+import java.util.*;
+
+@Aspect
+public class XRayInspector extends AbstractXRayInterceptor {
+    @Override
+    protected Map<String, Map<String, Object>> generateMetadata(
+            ProceedingJoinPoint proceedingJoinPoint,
+            Subsegment subsegment) {
+        return super.generateMetadata(proceedingJoinPoint, subsegment);
+    }
+
+    @Override
+    @Pointcut("@within(com.amazonaws.xray.spring.aop.XRayEnabled) && bean(*Controller)")
+    public void xrayEnabledClasses() {}
+}
