+++ src/main/resources/application-prod.properties
@@ -3,8 +3,6 @@
 # MIT license. See LICENSE file in root directory.
 #
 
-logging.level.com.amazonaws=TRACE
-
 spring.datasource.username=${SPRING_DATASOURCE_USERNAME}
 spring.datasource.password=${SPRING_DATASOURCE_PASSWORD}
 spring.datasource.url=${SPRING_DATASOURCE_URL}
